Stop scanning when crossing device boundaries, not mount points
The previous behavior was: if the EXCLUDE_MOUNTS flag is set, stop scanning
whenever a mount point is encountered, or any of the paths listed in the
'excluded-uris' setting; otherwise, scan the directory.

Now the scanner *always* excludes the locations listed in 'excluded-uris'.  If
the EXCLUDE_MOUNTS flag is set, it stops when a directory is on a different
unix device.  This is more consistent with other similar tools, and behaves
better in an OSTree system.